# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/app": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "新增AppInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth-token",
                        "name": "auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户名, 密码, 验证码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.AppInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回包括用户组信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AppInfo"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/app/{appId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "更新app信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth-token",
                        "name": "auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "app ID",
                        "name": "appId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.AppInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回AppInfo",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AppInfo"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "删除app版本信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth-token",
                        "name": "auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "App ID",
                        "name": "typeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdminType列表,返回包括列表,总数,页码,每页数量",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.QueryResult"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AppInfo": {
            "type": "object",
            "properties": {
                "appType": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deviceType": {
                    "type": "integer"
                },
                "forceUpdate": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "minVersionNum": {
                    "type": "string",
                    "example": "0"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "versionNum": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "response.QueryResult": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}